third_party/blink/web_tests/external/wpt/html/semantics/interactive-elements/the-dialog-element/dialog-focus-shadow-expected.txt

This is a testharness.js-based test.
FAIL show: No autofocus, no delegatesFocus, no siblings assert_equals: expected Element node <button tabindex="-1" id="focus-between-tests">Focus betw... but got Element node <div></div>
FAIL showModal: No autofocus, no delegatesFocus, no siblings assert_equals: expected Element node <body>

<!--
  We focus this one between each test, to en... but got Element node <div></div>
PASS show: No autofocus, no delegatesFocus, sibling before
PASS showModal: No autofocus, no delegatesFocus, sibling before
FAIL show: No autofocus, no delegatesFocus, sibling after assert_equals: expected Element node <button tabindex="-1" class="focus-me">Focusable</button> but got Element node <div></div>
FAIL showModal: No autofocus, no delegatesFocus, sibling after assert_equals: expected Element node <button tabindex="-1" class="focus-me">Focusable</button> but got Element node <div></div>
PASS show: No autofocus, yes delegatesFocus, no siblings
PASS showModal: No autofocus, yes delegatesFocus, no siblings
PASS show: No autofocus, yes delegatesFocus, sibling before
PASS showModal: No autofocus, yes delegatesFocus, sibling before
PASS show: No autofocus, yes delegatesFocus, sibling after
PASS showModal: No autofocus, yes delegatesFocus, sibling after
PASS show: Autofocus before, no delegatesFocus
PASS showModal: Autofocus before, no delegatesFocus
PASS show: Autofocus before, yes delegatesFocus
PASS showModal: Autofocus before, yes delegatesFocus
PASS show: Autofocus after, no delegatesFocus
PASS showModal: Autofocus after, no delegatesFocus
PASS show: Autofocus after, yes delegatesFocus
PASS showModal: Autofocus after, yes delegatesFocus
PASS show: Autofocus on shadow host, yes delegatesFocus, no siblings
PASS showModal: Autofocus on shadow host, yes delegatesFocus, no siblings
FAIL show: Autofocus on shadow host, yes delegatesFocus, sibling before assert_equals: expected Element node <div autofocus=""></div> but got Element node <button tabindex="-1">Focusable</button>
FAIL showModal: Autofocus on shadow host, yes delegatesFocus, sibling before assert_equals: expected Element node <div autofocus=""></div> but got Element node <button tabindex="-1">Focusable</button>
PASS show: Autofocus on shadow host, yes delegatesFocus, sibling after
PASS showModal: Autofocus on shadow host, yes delegatesFocus, sibling after
FAIL show: Autofocus on shadow host, no delegatesFocus, no siblings assert_equals: expected Element node <button tabindex="-1" id="focus-between-tests">Focus betw... but got Element node <div autofocus=""></div>
FAIL showModal: Autofocus on shadow host, no delegatesFocus, no siblings assert_equals: expected Element node <body>

<!--
  We focus this one between each test, to en... but got Element node <div autofocus=""></div>
PASS show: Autofocus on shadow host, no delegatesFocus, sibling before
PASS showModal: Autofocus on shadow host, no delegatesFocus, sibling before
FAIL show: Autofocus on shadow host, no delegatesFocus, sibling after assert_equals: expected Element node <button tabindex="-1" class="focus-me">Focusable</button> but got Element node <div autofocus=""></div>
FAIL showModal: Autofocus on shadow host, no delegatesFocus, sibling after assert_equals: expected Element node <button tabindex="-1" class="focus-me">Focusable</button> but got Element node <div autofocus=""></div>
PASS show: Autofocus inside shadow tree, yes delegatesFocus, no siblings
PASS showModal: Autofocus inside shadow tree, yes delegatesFocus, no siblings
FAIL show: Autofocus inside shadow tree, yes delegatesFocus, sibling before assert_equals: expected Element node <button tabindex="-1" class="focus-me">Focusable</button> but got Element node <div></div>
FAIL showModal: Autofocus inside shadow tree, yes delegatesFocus, sibling before assert_equals: expected Element node <button tabindex="-1" class="focus-me">Focusable</button> but got Element node <div></div>
FAIL show: Autofocus inside shadow tree, yes delegatesFocus, sibling after assert_equals: expected Element node <button tabindex="-1" class="focus-me">Focusable</button> but got Element node <div></div>
FAIL showModal: Autofocus inside shadow tree, yes delegatesFocus, sibling after assert_equals: expected Element node <button tabindex="-1" class="focus-me">Focusable</button> but got Element node <div></div>
FAIL show: Autofocus inside shadow tree, no delegatesFocus, no siblings assert_equals: expected Element node <button tabindex="-1" id="focus-between-tests">Focus betw... but got Element node <div></div>
FAIL showModal: Autofocus inside shadow tree, no delegatesFocus, no siblings assert_equals: expected Element node <body>

<!--
  We focus this one between each test, to en... but got Element node <div></div>
FAIL show: Autofocus inside shadow tree, no delegatesFocus, sibling before assert_equals: expected Element node <button tabindex="-1" class="focus-me">Focusable</button> but got Element node <div></div>
FAIL showModal: Autofocus inside shadow tree, no delegatesFocus, sibling before assert_equals: expected Element node <button tabindex="-1" class="focus-me">Focusable</button> but got Element node <div></div>
FAIL show: Autofocus inside shadow tree, no delegatesFocus, sibling after assert_equals: expected Element node <button tabindex="-1" class="focus-me">Focusable</button> but got Element node <div></div>
FAIL showModal: Autofocus inside shadow tree, no delegatesFocus, sibling after assert_equals: expected Element node <button tabindex="-1" class="focus-me">Focusable</button> but got Element node <div></div>
Harness: the test ran to completion.
